# api/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "오픈마켓 API",
    "description": "오픈마켓 API Server입니다.<br><a href=\"/\">버전별 변경사항 확인</a>"
  },
  "servers": [
    {
      "url": "https://market-lion.koyeb.app/api",
      "description": ""
    },
    {
      "url": "http://localhost/api",
      "description": "로컬 테스트"
    }
  ],
...
  "paths": {
...
    "/products/": {
      "get": {
        "tags": [
          "상품"
        ],
        "summary": "상품 목록 조회",
        "description": "상품 목록을 조회한다.",
        "parameters": [
          {
            "name": "minPrice",
            "description": "최저 가격",
            "in": "query",
            "default": 0,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "description": "최고 가격",
            "in": "query",
            "default": 99999999999,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "minShippingFees",
            "description": "최저 배송비",
            "in": "query",
            "default": 0,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxShippingFees",
            "description": "최고 배송비",
            "in": "query",
            "default": 99999999999,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "description": "상품명 검색어",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seller_id",
            "description": "판매자 id",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "custom",
            "description": "custom 데이터 예시: {&quot;extra.isNew&quot;: true}",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/productListRes"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        }
      }
    },
    "/products/{_id}": {
      "get": {
        "tags": [
          "상품"
        ],
        "summary": "상품 상세 조회",
        "description": "상품 상세 정보를 조회한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "상품 id",
            "example": 4
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/productInfoRes"
                }
              }
            }
          },
          "404": {
            "description": "리소스가 존재하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        }
      }
    },
...
    "/orders/{_id}/products/{product_id}": {
      "patch": {
        "tags": [
          "구매"
        ],
        "summary": "상품별 주문 상태 수정",
        "description": "상품별로 주문 상태를 수정한다.",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
...
    "/replies/products/{_id}": {
      "get": {
        "tags": [
          "구매 후기"
        ],
        "summary": "상품 구매 후기 목록",
        "description": "지정한 상품의 구매 후기 목록을 조회한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
...
    "/replies/seller/{seller_id}": {
      "get": {
        "tags": [
          "구매 후기"
        ],
        "summary": "판매자 구매 후기 목록",
        "description": "판매자의 상품에 등록된 모든 구매 후기 목록을 조회한다.",
        "parameters": [
          {
            "name": "seller_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
...
    "/bookmarks/products/{product_id}": {
      "get": {
        "tags": [
          "북마크"
        ],
        "summary": "북마크 한건 조회",
        "description": "지정한 상품에 대한 나의 북마크 정보를 조회한다.",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
...
    "/posts/seller/{_id}": {
      "get": {
        "tags": [
          "게시판"
        ],
        "summary": "판매자 게시물 목록",
        "description": "판매자의 상품에 등록된 게시물 목록을 조회한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
...
    "/seller/products/": {
      "get": {
        "tags": [
          "상품 관리"
        ],
        "summary": "판매 상품 목록 조회",
        "description": "자신의 판매 상품 목록을 조회한다.",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      },
      "post": {
        "tags": [
          "상품 관리"
        ],
        "summary": "상품 등록",
        "description": "상품을 등록합니다.<br>상품 등록 후 상품 정보를 반환합니다.",
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/productCreateRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "422": {
            "description": "입력값 검증 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error422"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "상품 정보",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/productCreate"
              }
            }
          }
        }
      }
    },
    "/seller/products/{_id}": {
      "get": {
        "tags": [
          "상품 관리"
        ],
        "summary": "판매 상품 상세 조회",
        "description": "자신의 판매 상품 정보를 조회한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "상품 id",
            "example": 6
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/productInfoRes"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      },
      "patch": {
        "tags": [
          "상품 관리"
        ],
        "summary": "상품 수정",
        "description": "상품을 수정한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "상품 id",
            "example": "6"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/productUpdateRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "상품이 존재하지 않거나 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "422": {
            "description": "입력값 검증 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error422"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "수정할 상품 정보",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/productUpdate"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "상품 관리"
        ],
        "summary": "상품 삭제",
        "description": "상품을 삭제한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "상품 id",
            "example": "6"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/simpleOK"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "상품이 존재하지 않거나 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
...
    "/seller/orders/{_id}/products/{product_id}": {
      "patch": {
        "tags": [
          "주문 관리"
        ],
        "summary": "상품별 주문 상태 수정",
        "description": "상품 단위로 주문 상태를 수정한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Access Token": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      },
      "Refresh Token": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "error401": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "{인증 실패 사유}"
          },
          "errorName": {
            "type": "string",
            "example": "EmptyAuthorization | TokenExpiredError | JsonWebTokenError"
          }
        }
      },
...
      "error404": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "/api/xxx 리소스를 찾을 수 없습니다."
          }
        }
      },
...
      "error500": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "요청하신 작업 처리에 실패했습니다. 잠시 후 다시 이용해 주시기 바랍니다."
          }
        }
      },
      "simpleOK": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 1
          }
        }
      },
...
      "productCreateRes": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 1
          },
          "item": {
            "type": "object",
            "properties": {
              "price": {
                "type": "number",
                "example": 22000
              },
              "shippingFees": {
                "type": "number",
                "example": 3000
              },
              "show": {
                "type": "boolean",
                "example": true
              },
              "active": {
                "type": "boolean",
                "example": true
              },
              "name": {
                "type": "string",
                "example": "ZOZOFO 테이블 게임 축구 보드 사커 게임기 보드게임 2인경기 완구 가족모임 미니 월드컵 스포츠 어린이 크리스마스 선물 생일 선물"
              },
              "mainImages": {
                "type": "array",
                "example": [
                  "/uploads/sample-janngu.jpg"
                ],
                "items": {
                  "type": "string"
                }
              },
              "content": {
                "type": "string",
                "example": "<div class=\"product-detail\"><p>ZOZOFO 테이블 게임 축구 보드 사커 게임기 보드게임 2인경기 완구 가족모임 미니 월드컵 스포츠 어린이 크리스마스 선물 생일 선물 상세 설명</p></div>"
              },
              "createdAt": {
                "type": "string",
                "example": "2023.11.22 07:26:35"
              },
              "updatedAt": {
                "type": "string",
                "example": "2023.11.22 07:26:35"
              },
              "extra": {
                "type": "object",
                "properties": {
                  "isNew": {
                    "type": "boolean",
                    "example": true
                  },
                  "isBest": {
                    "type": "boolean",
                    "example": true
                  },
                  "category": {
                    "type": "array",
                    "example": [
                      "PC02",
                      "PC0201"
                    ],
                    "items": {
                      "type": "string"
                    }
                  },
                  "quantity": {
                    "type": "number",
                    "example": 600
                  },
                  "buyQuantity": {
                    "type": "number",
                    "example": 190
                  },
                  "order": {
                    "type": "number",
                    "example": 7
                  }
                }
              },
              "seller_id": {
                "type": "number",
                "example": 2
              },
              "_id": {
                "type": "number",
                "example": 15
              }
            }
          }
        }
      },
      "productListRes": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 1
          },
          "item": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "number",
                  "example": 12
                },
                "seller_id": {
                  "type": "number",
                  "example": 2
                },
                "price": {
                  "type": "number",
                  "example": 9000
                },
                "shippingFees": {
                  "type": "number",
                  "example": 3000
                },
                "show": {
                  "type": "boolean",
                  "example": true
                },
                "active": {
                  "type": "boolean",
                  "example": true
                },
                "name": {
                  "type": "string",
                  "example": "스키비디 토일렛 봉제 인형 (25cm-30cm) 시리즈 크리스마스 선물"
                },
                "mainImages": {
                  "type": "array",
                  "example": [
                    "/uploads/sample-skibidi11.jpg"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "createdAt": {
                  "type": "string",
                  "example": "2023.11.10 07:07:41"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2023.11.16 13:07:41"
                },
                "extra": {
                  "type": "object",
                  "properties": {
                    "isNew": {
                      "type": "boolean",
                      "example": true
                    },
                    "isBest": {
                      "type": "boolean",
                      "example": true
                    },
                    "category": {
                      "type": "array",
                      "example": [
                        "PC01",
                        "PC0103"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "quantity": {
                      "type": "number",
                      "example": 999
                    },
                    "buyQuantity": {
                      "type": "number",
                      "example": 230
                    },
                    "order": {
                      "type": "number",
                      "example": 7
                    }
                  }
                }
              }
            }
          }
        }
      },
      "productInfoRes": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 1
          },
          "item": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "number",
                "example": 4
              },
              "seller_id": {
                "type": "number",
                "example": 3
              },
              "price": {
                "type": "number",
                "example": 45000
              },
              "shippingFees": {
                "type": "number",
                "example": 3500
              },
              "show": {
                "type": "boolean",
                "example": true
              },
              "active": {
                "type": "boolean",
                "example": true
              },
              "name": {
                "type": "string",
                "example": "레고 테크닉 42151 부가티 볼리드"
              },
              "mainImages": {
                "type": "array",
                "example": [
                  "/uploads/sample-bugatti.png"
                ],
                "items": {
                  "type": "string"
                }
              },
              "content": {
                "type": "string",
                "example": "\n        <div class=\"product-detail\">\n          <p>레고 테크닉 42151 부가티 볼리드 상세 설명</p>\n        </div>"
              },
              "createdAt": {
                "type": "string",
                "example": "2023.10.19 12:07:41"
              },
              "updatedAt": {
                "type": "string",
                "example": "2023.10.30 16:07:41"
              },
              "extra": {
                "type": "object",
                "properties": {
                  "isNew": {
                    "type": "boolean",
                    "example": false
                  },
                  "isBest": {
                    "type": "boolean",
                    "example": true
                  },
                  "category": {
                    "type": "array",
                    "example": [
                      "PC03",
                      "PC0303"
                    ],
                    "items": {
                      "type": "string"
                    }
                  },
                  "quantity": {
                    "type": "number",
                    "example": 100
                  },
                  "buyQuantity": {
                    "type": "number",
                    "example": 30
                  },
                  "order": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "replies": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "number",
                      "example": 2
                    },
                    "rating": {
                      "type": "number",
                      "example": 4
                    },
                    "content": {
                      "type": "string",
                      "example": "배송이 좀 느려요."
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2023.11.18 18:07:41"
                    },
                    "userName": {
                      "type": "string",
                      "example": "네오"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
...
  }
}